[HVM-261] vmdhcpd spinning at 100% CPU after bridge is removed
The problem here turned out to be node_pcap: if an interface is
unplumbed out from under it, its fd keeps on reporting that data is
available to read.  Its IOWatcher then continuously spins on that data
being available, even though there is none.  The solution to this is to
detect whether the interface is still available on empty reads.  If it's
not, then close the pcap session.

Also, don't do an update as part of the default make target, since it's
really annoying when developing.