Remove package-lock.json
For some reason, this file is causing `make check` to fail. After
removing this file `make check` passes, and subsequently running
`npm install` does not create this file. Given that, it seems it
doesn't need to exist, and because it's causing a problem I'm just
going to nuke it outright.